fix(toast): Move useToastCloseButton to toastHooks
